+++ src/calibrate/requirements.txt
@@ -11,4 +11,5 @@ pymongo==3.11.3
 python-dotenv==0.17.0
 scikit_learn==0.24.1
 confluent-kafka
-confluent-kafka[avro]
+confluent-avro
+kafka-python
